requirements.txt
flask
google-generativeai
numpy
python-dotenv
matplotlib
logging
glob
datetime
json